Refactoring package; support unload & load at startup
