+++ sql12/plugins/nimbusdark/src/net/sourceforge/squirrel_sql/plugin/nimbusdark/close.gif
